+++ changelog.md
@@ -1,3 +1,17 @@
+**V1.2.4**
+- Added webhook.php to excludes https://github.com/DonutsNL/samlsso/issues/32
+- Added `.php` to excludes path in `Config.php` to make sure the 'add' button shows.
+- Altered `getConfigIdByEmailDomain()` to handle multiple domains per IDP using comma seperated domain lists i.e. `domain1.com,domain1.nl,domain1.org`
+- Altered the description of the userDomain field explaining the multiple domain option.
+- Added basics in `performSamlSSO()` and `doAuth()` for Subject hinting, but this isnt supported by Entra. Maybe make this a configurable option in the future for IDPs that do support Subject hinting.
+- Add acs exclusion in doAuth flow https://github.com/DonutsNL/samlsso/issues/29
+- Remove excessive state inits in authFlow object.
+- Fix issue in logout method causing warnings to be logged by Html object. https://github.com/DonutsNL/samlsso/issues/35
+- Updated the credits in the Readme.md (long overdue sorry for that)
+- Updated all the translation projects in https://app.transifex.com/quinquies/glpisaml/
+- Added /api.php to excludes as requested by https://github.com/DonutsNL/samlsso/issues/36
+
+
 **V1.2.2**
 - Corrected a router bug that caused the http header bag and response objects to be called with __toString()
 - Added an additional Prerequisites check to validate the php cookie settings and block install if they are not correct (issue 13).